elsa2/requirements.txt
Flask==1.1.1
Flask-WTF==0.14.3
PyJWT==1.7.1
pytest==6.1.1
WTForms==2.3.3
